(PC-21951)[PRO] refactor: move down in the component infos from reimbursements/details route
