+++ pro/src/pages/Reimbursements/Reimbursement.scss
@@ -14,6 +14,10 @@
     }
   }
 
+  .spinner {
+    margin-top: rem.torem(36px);
+  }
+
   .reset-filters {
     @include fonts.button;
