+++ pro/src/pages/Reimbursements/ReimbursementsDetails/ReimbursementsDetails.tsx
@@ -1,9 +1,13 @@
-import React, { useEffect, useState } from 'react'
+import React, { useCallback, useEffect, useState } from 'react'
 
+import { VenueListItemResponseModel } from 'apiClient/v1'
 import ButtonDownloadCSV from 'components/ButtonDownloadCSV'
+import getVenuesForOffererAdapter from 'core/Venue/adapters/getVenuesForOffererAdapter'
 import useCurrentUser from 'hooks/useCurrentUser'
 import { ButtonLink } from 'ui-kit/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
+import Icon from 'ui-kit/Icon/Icon'
+import Spinner from 'ui-kit/Spinner/Spinner'
 import { API_URL } from 'utils/config'
 import {
   FORMAT_ISO_DATE_ONLY,
@@ -11,18 +15,13 @@ import {
   getToday,
 } from 'utils/date'
 import { stringify } from 'utils/query-string'
+import { sortByDisplayName } from 'utils/strings'
 
 import DetailsFilters from './DetailsFilters'
 
-type venuesOptionsType = [
-  {
-    id: string
-    displayName: string
-  }
-]
-
-interface IReimbursementsDetailsProps {
-  venuesOptions: venuesOptionsType
+type VenueOptionType = {
+  id: string
+  displayName: string
 }
 
 interface ICsvQueryParams {
@@ -31,9 +30,7 @@ interface ICsvQueryParams {
   reimbursementPeriodEndingDate?: string
 }
 
-const ReimbursementsDetails = ({
-  venuesOptions,
-}: IReimbursementsDetailsProps): JSX.Element => {
+const ReimbursementsDetails = (): JSX.Element => {
   const { currentUser } = useCurrentUser()
   const ALL_VENUES_OPTION_ID = 'allVenues'
   const today = getToday()
@@ -51,6 +48,8 @@ const ReimbursementsDetails = ({
   const [filters, setFilters] = useState(INITIAL_FILTERS)
   const { venue, periodStart, periodEnd } = filters
   const [csvQueryParams, setCsvQueryParams] = useState('')
+  const [venuesOptions, setVenuesOptions] = useState<VenueOptionType[]>([])
+  const [isLoading, setIsLoading] = useState(true)
 
   const dateFilterFormat = (date: Date) =>
     formatBrowserTimezonedDateAsUTC(date, FORMAT_ISO_DATE_ONLY)
@@ -59,6 +58,32 @@ const ReimbursementsDetails = ({
     currentUser.isAdmin && venue === ALL_VENUES_OPTION_ID
   const shouldDisableButtons =
     !isPeriodFilterSelected || requireVenueFilterForAdmin
+
+  const buildAndSortVenueFilterOptions = (
+    venues: VenueListItemResponseModel[]
+  ) =>
+    sortByDisplayName(
+      venues.map(venue => ({
+        id: venue.id,
+        displayName: venue.isVirtual
+          ? `${venue.offererName} - Offre numérique`
+          : /* istanbul ignore next: TO FIX */
+            venue.publicName || venue.name,
+      }))
+    )
+
+  const loadVenues = useCallback(async () => {
+    const venuesResponse = await getVenuesForOffererAdapter({
+      activeOfferersOnly: true,
+    })
+    const selectOptions = buildAndSortVenueFilterOptions(venuesResponse.payload)
+    setVenuesOptions(selectOptions)
+    setIsLoading(false)
+  }, [setVenuesOptions])
+
+  useEffect(() => {
+    loadVenues()
+  }, [loadVenues])
 
   useEffect(() => {
     const params: ICsvQueryParams = {}
@@ -73,7 +98,21 @@ const ReimbursementsDetails = ({
     }
     setCsvQueryParams(stringify(params))
   }, [periodEnd, periodStart, venue])
-
+  if (isLoading) {
+    return (
+      <div className="spinner">
+        <Spinner />
+      </div>
+    )
+  }
+  if (venuesOptions.length === 0) {
+    return (
+      <div className="no-refunds">
+        <Icon alt="" svg="ico-no-bookings" />
+        <span>Aucun remboursement à afficher</span>
+      </div>
+    )
+  }
   return (
     <>
       <DetailsFilters
